refactor: move controller package
