+++ src/test/java/controller/StaticFileControllerTest.java
@@ -1,11 +1,9 @@
-package http.controller;
+package controller;
 
-import db.DataBase;
 import http.request.HttpRequest;
 import http.request.RequestHandler;
 import http.response.HttpResponse;
 import http.response.ResponseHandler;
-import model.User;
 import org.junit.jupiter.api.Test;
 
 import java.io.*;
@@ -13,28 +11,24 @@ import java.nio.charset.StandardCharsets;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class LoginControllerTest {
+class StaticFileControllerTest {
     private static final String TEST_DIRECTORY = "./src/test/resources/";
 
     private HttpRequest httpRequest;
     private HttpResponse httpResponse;
 
     @Test
-    public void doPostTest() throws Exception {
-        InputStream in = new FileInputStream(new File(TEST_DIRECTORY + "Http_POST_Login.txt"));
+    public void doGetTest() throws Exception {
+        InputStream in = new FileInputStream(new File(TEST_DIRECTORY + "Http_GET_File.txt"));
         BufferedReader br = new BufferedReader(new InputStreamReader(in, StandardCharsets.UTF_8));
         httpRequest = new RequestHandler(br).create();
         httpResponse = new ResponseHandler().create();
         httpResponse.addHeaderFromRequest(httpRequest);
 
-        User user = new User("cony", "password", "cony", "cony@cony.com");
-        DataBase.addUser(user);
+        StaticFileController staticFileController = new StaticFileController();
+        staticFileController.doGet(httpRequest, httpResponse);
 
-        LoginController loginController = new LoginController();
-        loginController.doPost(httpRequest, httpResponse);
-
-        assertThat(httpResponse.toString()).contains("302 FOUND");
-        assertThat(httpResponse.toString()).contains("Location: /index.html");
-        assertThat(httpResponse.toString()).contains("JSESSIONID");
+        assertThat(httpResponse.toString()).contains("200 OK");
+        assertThat(httpResponse.toString()).contains("Content-Length: 7049");
     }
 }
